+++ quickwit/quickwit-actors/src/lib.rs
@@ -69,7 +69,7 @@ pub use self::actor_state::ActorState;
 pub use self::channel_with_priority::{QueueCapacity, RecvError, SendError, TrySendError};
 pub use self::mailbox::{Inbox, Mailbox};
 pub use self::registry::ActorObservation;
-pub use self::supervisor::{Supervisor, SupervisorState};
+pub use self::supervisor::{Supervisor, SupervisorMetrics, SupervisorState};
 
 /// Heartbeat used to verify that actors are progressing.
 ///
